python_scraping/requirements.txt
aiohappyeyeballs==2.3.5
aiohttp==3.10.3
aiosignal==1.3.1
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cloudevents==1.11.0
deprecation==2.1.0
dnspython==2.6.1
Flask==3.0.3
frozenlist==1.4.1
functions-framework==3.8.1
gunicorn==23.0.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
multidict==6.0.5
mysql-connector-python==9.0.0
mysqlclient==2.2.4
packaging==24.1
pymongo==4.8.0
python-dotenv==1.0.1
requests==2.32.3
soupsieve==2.5
urllib3==2.2.2
watchdog==4.0.2
Werkzeug==3.0.3
yarl==1.9.4
